requirements.txt
Flask==2.3.2
torch==2.0.1
transformers==4.30.0
requests==2.31.0
python-dotenv==1.0.0